separate tests into classes specific to importer type
